[Exception] Move ExceptionHandler to new Handler namespace.
