+++ src/Valkyrja/Application/Entry/App.php
@@ -23,8 +23,8 @@ use Valkyrja\Cli\Interaction\Factory\InputFactory;
 use Valkyrja\Cli\Interaction\Input\Contract\Input;
 use Valkyrja\Cli\Server\Handler\Contract\InputHandler;
 use Valkyrja\Container\Manager\Contract\Container;
-use Valkyrja\Exception\Contract\ExceptionHandler as ErrorHandlerContract;
-use Valkyrja\Exception\ExceptionHandler;
+use Valkyrja\Exception\Handler\Contract\ExceptionHandler as ErrorHandlerContract;
+use Valkyrja\Exception\Handler\ExceptionHandler;
 use Valkyrja\Http\Message\Factory\RequestFactory;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Http\Server\Handler\Contract\RequestHandler;
