+++ src/Valkyrja/Exception/Handler/Contract/ExceptionHandler.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Exception\Contract;
+namespace Valkyrja\Exception\Handler\Contract;
 
 use Throwable;
