+++ src/Valkyrja/Http/Routing/Controller/ApiController.php
@@ -16,7 +16,7 @@ namespace Valkyrja\Http\Routing\Controller;
 use Throwable;
 use Valkyrja\Api\Constant\Status;
 use Valkyrja\Api\Manager\Contract\Api;
-use Valkyrja\Exception\ExceptionHandler;
+use Valkyrja\Exception\Handler\ExceptionHandler;
 use Valkyrja\Http\Message\Enum\StatusCode;
 use Valkyrja\Http\Message\Factory\Contract\ResponseFactory;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
